server: sid is now written as a byte

+++ BombermanServer/BombermanServer.cs
@@ -75,7 +75,7 @@ namespace BombermanServer
 
                                 NetOutgoingMessage outmsg = server.CreateMessage();
                                 // senderID, PacketType, ID
-                                outmsg.Write(0);
+                                outmsg.Write((byte)0);
                                 outmsg.Write((byte)PacketTypeEnums.PacketType.SEND_PLAYER_ID);
                                 outmsg.Write((byte)playersConnected);
                                 // we don't want this to be lost, so set level to ReliableOrdered
@@ -93,7 +93,7 @@ namespace BombermanServer
                                 {
                                     Console.WriteLine($"Broadcasting to connected player {i + 1} that new player {playersConnected} has connected");
                                     NetOutgoingMessage newPlayerMsg = server.CreateMessage();
-                                    newPlayerMsg.Write(0);
+                                    newPlayerMsg.Write((byte)0);
                                     newPlayerMsg.Write((byte)PacketTypeEnums.PacketType.NEW_PLAYER_ID);
                                     newPlayerMsg.Write((byte)playersConnected);
                                     server.SendMessage(newPlayerMsg, playerInfoArr[i].playerConnection, NetDeliveryMethod.ReliableOrdered, 0);
@@ -133,7 +133,7 @@ namespace BombermanServer
             {
                 Console.WriteLine($"Broadcasting to connected player {i + 1} that game has started");
                 NetOutgoingMessage newPlayerMsg = server.CreateMessage();
-                newPlayerMsg.Write(0);
+                newPlayerMsg.Write((byte)0);
                 newPlayerMsg.Write((byte)PacketTypeEnums.PacketType.GAME_START);
                 server.SendMessage(newPlayerMsg, playerInfoArr[i].playerConnection, NetDeliveryMethod.ReliableOrdered, 0);
             }
